+++ src/routes.js
@@ -3,10 +3,11 @@
 
     var module = angular.module('angular-token-auth.routes', [
         'angular-token-auth.constants',
-        'project_settings'
+        'project_settings',
+        'ngRoute'
     ]);
 
-    module.config(['$routeProvider','TOKEN_AUTH', 'PROJECT_SETTINGS', function ($routeProvider, TOKEN_AUTH, PROJECT_SETTINGS) {
+    module.config(['$routeProvider', 'TOKEN_AUTH', 'PROJECT_SETTINGS', function ($routeProvider, TOKEN_AUTH, PROJECT_SETTINGS) {
         var MODULE_SETTINGS = angular.extend({}, TOKEN_AUTH, PROJECT_SETTINGS.TOKEN_AUTH);
 
         $routeProvider